requirements.txt
asgiref==3.4.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-bootstrap4==3.0.1
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pillow==8.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
